LPS-51648 Fix jsp compilation
