+++ modules/apps/bookmarks/bookmarks-service/src/com/liferay/bookmarks/util/BookmarksUtil.java
@@ -15,18 +15,10 @@
 package com.liferay.bookmarks.util;
 
 import com.liferay.bookmarks.constants.BookmarksPortletKeys;
-import com.liferay.bookmarks.model.BookmarksEntry;
 import com.liferay.bookmarks.model.BookmarksFolder;
 import com.liferay.bookmarks.model.BookmarksFolderConstants;
-import com.liferay.bookmarks.service.BookmarksEntryLocalServiceUtil;
 import com.liferay.bookmarks.service.BookmarksFolderLocalServiceUtil;
 import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.search.Document;
-import com.liferay.portal.kernel.search.Field;
-import com.liferay.portal.kernel.search.Hits;
-import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.theme.ThemeDisplay;
@@ -34,7 +26,6 @@ import com.liferay.portal.util.PortalUtil;
 import com.liferay.portal.util.WebKeys;
 import com.liferay.portlet.PortletURLFactoryUtil;
 
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
@@ -99,44 +90,5 @@ public class BookmarksUtil {
 
 		return portletURL.toString();
 	}
-
-	public static List<Object> getEntries(Hits hits) {
-		List<Object> entries = new ArrayList<Object>();
-
-		for (Document document : hits.getDocs()) {
-			String entryClassName = document.get(Field.ENTRY_CLASS_NAME);
-			long entryClassPK = GetterUtil.getLong(
-				document.get(Field.ENTRY_CLASS_PK));
-
-			try {
-				Object obj = null;
-
-				if (entryClassName.equals(BookmarksEntry.class.getName())) {
-					obj = BookmarksEntryLocalServiceUtil.getEntry(entryClassPK);
-				}
-				else if (entryClassName.equals(
-							BookmarksFolder.class.getName())) {
-
-					obj = BookmarksFolderLocalServiceUtil.getFolder(
-						entryClassPK);
-				}
-
-				entries.add(obj);
-			}
-			catch (Exception e) {
-				if (_log.isWarnEnabled()) {
-					_log.warn(
-						"Bookmarks search index is stale and contains entry " +
-							entryClassPK);
-				}
-
-				continue;
-			}
-		}
-
-		return entries;
-	}
-
-	private static final Log _log = LogFactoryUtil.getLog(BookmarksUtil.class);
 
 }
