+++ modules/apps/bookmarks/bookmarks-web/docroot/WEB-INF/src/com/liferay/bookmarks/web/portlet/util/BookmarksUtil.java
@@ -17,6 +17,7 @@ package com.liferay.bookmarks.web.portlet.util;
 import com.liferay.bookmarks.model.BookmarksEntry;
 import com.liferay.bookmarks.model.BookmarksFolder;
 import com.liferay.bookmarks.model.BookmarksFolderConstants;
+import com.liferay.bookmarks.service.BookmarksEntryLocalServiceUtil;
 import com.liferay.bookmarks.service.BookmarksFolderLocalServiceUtil;
 import com.liferay.bookmarks.util.comparator.EntryCreateDateComparator;
 import com.liferay.bookmarks.util.comparator.EntryModifiedDateComparator;
@@ -25,7 +26,13 @@ import com.liferay.bookmarks.util.comparator.EntryPriorityComparator;
 import com.liferay.bookmarks.util.comparator.EntryURLComparator;
 import com.liferay.bookmarks.util.comparator.EntryVisitsComparator;
 import com.liferay.portal.kernel.language.LanguageUtil;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.portlet.LiferayWindowState;
+import com.liferay.portal.kernel.search.Document;
+import com.liferay.portal.kernel.search.Field;
+import com.liferay.portal.kernel.search.Hits;
+import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.HtmlUtil;
 import com.liferay.portal.kernel.util.OrderByComparator;
 import com.liferay.portal.kernel.util.ParamUtil;
@@ -34,6 +41,7 @@ import com.liferay.portal.model.Company;
 import com.liferay.portal.theme.ThemeDisplay;
 import com.liferay.portal.util.PortalUtil;
 
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -185,6 +193,43 @@ public class BookmarksUtil {
 		return definitionTerms;
 	}
 
+	public static List<Object> getEntries(Hits hits) {
+		List<Object> entries = new ArrayList<Object>();
+
+		for (Document document : hits.getDocs()) {
+			String entryClassName = document.get(Field.ENTRY_CLASS_NAME);
+			long entryClassPK = GetterUtil.getLong(
+				document.get(Field.ENTRY_CLASS_PK));
+
+			try {
+				Object obj = null;
+
+				if (entryClassName.equals(BookmarksEntry.class.getName())) {
+					obj = BookmarksEntryLocalServiceUtil.getEntry(entryClassPK);
+				}
+				else if (entryClassName.equals(
+							BookmarksFolder.class.getName())) {
+
+					obj = BookmarksFolderLocalServiceUtil.getFolder(
+						entryClassPK);
+				}
+
+				entries.add(obj);
+			}
+			catch (Exception e) {
+				if (_log.isWarnEnabled()) {
+					_log.warn(
+						"Bookmarks search index is stale and contains entry " +
+							entryClassPK);
+				}
+
+				continue;
+			}
+		}
+
+		return entries;
+	}
+
 	public static OrderByComparator<BookmarksEntry> getEntryOrderByComparator(
 		String orderByCol, String orderByType) {
 
@@ -217,5 +262,7 @@ public class BookmarksUtil {
 
 		return orderByComparator;
 	}
+
+	private static final Log _log = LogFactoryUtil.getLog(BookmarksUtil.class);
 
 }
